moved downloader to files/internal/downloader
