+++ cli/cli.go
@@ -262,7 +262,6 @@ func newCommandTree(ru *run.Run) (rootCmd *cobra.Command) {
 	addCmd(ru, xCmd, newXLockSrcCmd())
 	addCmd(ru, xCmd, newXLockConfigCmd())
 	addCmd(ru, xCmd, newXProgressCmd())
-	addCmd(ru, xCmd, newXDownloadCmd())
 
 	return rootCmd
 }
