+++ libsq/files/internal/downloader/LICENSE.txt
